cmake: allow tests to be disabled via BUILD_TESTING
Testing is broken since it relies on things like
CMAKE_Fortran_IMPLICIT_LINK_LIBRARIES and similar things for HIP and CUDA that
don't work unless you explicitly enable the necessary language support in
CMake. This commit allows to disable the tests for installation via Spack while
this is being fixed.